+++ triple_arena/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changes
 - Added `Send` and `Sync` bounds to `PtrTrait`
+- Added an argument to `create` in `try_insert_with`
 
 ### Additions
 - Added Changelog
+- Added `insert_with`